requirements.txt
colorama==0.4.6
pygame-ce==2.3.2